refactor: unify project link stuff
