+++ pages/dataset/[datasetId]/[datasetVersion].tsx
@@ -4,11 +4,11 @@ import type { GetServerSideProps } from "next";
 import NextError from "next/error";
 import { useRouter } from "next/router";
 
-import { API_ROUTES } from "../../../constants/routes";
 import { PlaintextViewer } from "../../../features/PlaintextViewer";
 import type { NotSuccessful, Successful } from "../../../utils/api/plaintextViewerSSR";
 import { plaintextViewerSSR } from "../../../utils/api/plaintextViewerSSR";
 import { createErrorProps } from "../../../utils/api/serverSidePropsError";
+import { API_ROUTES } from "../../../utils/app/routes";
 
 export type DatasetVersionProps = Successful | NotSuccessful;
